remove float64 in get_transfer_index.

+++ llada/decoding/utils.py
@@ -62,7 +62,7 @@ def get_transfer_index(logits, temperature, remasking, mask_index, x, num_transf
     x0 = torch.argmax(logits_with_noise, dim=-1) # b, l
 
     if remasking == 'low_confidence':
-        p = F.softmax(logits.to(torch.float64), dim=-1)
+        p = F.softmax(logits, dim=-1)
         x0_p = torch.squeeze(
             torch.gather(p, dim=-1, index=torch.unsqueeze(x0, -1)), -1) # b, l
     elif remasking == 'random':
